swarm/pss: Separate test file proto+hs
