+++ swarm/pss/pss_test.go
@@ -324,234 +324,6 @@ func TestKeys(t *testing.T) {
 	//psp := ps.symKeyPool[inkeyid][topic]
 }
 
-// asymmetrical key exchange between two directly connected peers
-// full address, partial address (8 bytes) and empty address
-func TestHandshake(t *testing.T) {
-	t.Run("32", testHandshake)
-	t.Run("8", testHandshake)
-	t.Run("0", testHandshake)
-}
-
-func testHandshake(t *testing.T) {
-
-	// how much of the address we will use
-	useHandshake = true
-	var addrsize int64
-	var err error
-	addrsizestring := strings.Split(t.Name(), "/")
-	addrsize, _ = strconv.ParseInt(addrsizestring[1], 10, 0)
-
-	// set up two nodes directly connected
-	// (we are not testing pss routing here)
-	topic := whisper.BytesToTopic([]byte("foo:42"))
-	hextopic := fmt.Sprintf("%x", topic)
-
-	clients, err := setupNetwork(2)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	var loaddr []byte
-	err = clients[0].Call(&loaddr, "pss_baseAddr")
-	if err != nil {
-		t.Fatalf("rpc get node 1 baseaddr fail: %v", err)
-	}
-	loaddr = loaddr[:addrsize]
-	var roaddr []byte
-	err = clients[1].Call(&roaddr, "pss_baseAddr")
-	if err != nil {
-		t.Fatalf("rpc get node 2 baseaddr fail: %v", err)
-	}
-	roaddr = roaddr[:addrsize]
-	log.Debug("addresses", "left", loaddr, "right", roaddr)
-
-	// retrieve public key from pss instance
-	// set this public key reciprocally
-	lpubkey := make([]byte, 32)
-	err = clients[0].Call(&lpubkey, "pss_getPublicKey")
-	if err != nil {
-		t.Fatalf("rpc get node 1 pubkey fail: %v", err)
-	}
-	rpubkey := make([]byte, 32)
-	err = clients[1].Call(&rpubkey, "pss_getPublicKey")
-	if err != nil {
-		t.Fatalf("rpc get node 2 pubkey fail: %v", err)
-	}
-
-	time.Sleep(time.Millisecond * 1000) // replace with hive healthy code
-
-	// give each node its peer's public key
-	err = clients[0].Call(nil, "pss_setPeerPublicKey", rpubkey, hextopic, roaddr)
-	if err != nil {
-		t.Fatal(err)
-	}
-	err = clients[1].Call(nil, "pss_setPeerPublicKey", lpubkey, hextopic, loaddr)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	// perform the handshake
-	// after this each side will have defaultSymKeyBufferCapacity symkeys each for in- and outgoing messages:
-	// L -> request 6 keys -> R
-	// L <- send 6 keys, request 6 keys <- R
-	// L -> send 6 keys -> R
-	// the call will fill the array with symkeys L needs for sending to R
-	err = clients[0].Call(nil, "pss_addHandshake", hextopic)
-	if err != nil {
-		t.Fatal(err)
-	}
-	err = clients[1].Call(nil, "pss_addHandshake", hextopic)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	var lhsendsymkeyids []string
-	err = clients[0].Call(&lhsendsymkeyids, "pss_handshake", common.ToHex(rpubkey), hextopic, roaddr, true)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	// make sure the r-node gets its keys
-	time.Sleep(time.Second)
-
-	// check if we have 6 outgoing keys stored, and they match what was received from R
-	var lsendsymkeyids []string
-	err = clients[0].Call(&lsendsymkeyids, "pss_getHandshakeKeys", common.ToHex(rpubkey), hextopic, false, true)
-	if err != nil {
-		t.Fatal(err)
-	}
-	m := 0
-	for _, hid := range lhsendsymkeyids {
-		for _, lid := range lsendsymkeyids {
-			if lid == hid {
-				m++
-			}
-		}
-	}
-	if m != defaultSymKeyCapacity {
-		t.Fatalf("buffer size mismatch, expected %d, have %d: %v", defaultSymKeyCapacity, m, lsendsymkeyids)
-	}
-
-	// check if in- and outgoing keys on l-node and r-node match up and are in opposite categories (l recv = r send, l send = r recv)
-	var rsendsymkeyids []string
-	err = clients[1].Call(&rsendsymkeyids, "pss_getHandshakeKeys", common.ToHex(lpubkey), hextopic, false, true)
-	if err != nil {
-		t.Fatal(err)
-	}
-	var lrecvsymkeyids []string
-	err = clients[1].Call(&lrecvsymkeyids, "pss_getHandshakeKeys", common.ToHex(rpubkey), hextopic, true, false)
-	if err != nil {
-		t.Fatal(err)
-	}
-	var rrecvsymkeyids []string
-	err = clients[1].Call(&rrecvsymkeyids, "pss_getHandshakeKeys", common.ToHex(lpubkey), hextopic, true, false)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	// get outgoing symkeys in byte form from both sides
-	var lsendsymkeys [][]byte
-	for _, id := range lsendsymkeyids {
-		var key []byte
-		err = clients[0].Call(&key, "pss_getSymmetricKey", id)
-		if err != nil {
-			t.Fatal(err)
-		}
-		lsendsymkeys = append(lsendsymkeys, key)
-	}
-	var rsendsymkeys [][]byte
-	for _, id := range rsendsymkeyids {
-		var key []byte
-		err = clients[1].Call(&key, "pss_getSymmetricKey", id)
-		if err != nil {
-			t.Fatal(err)
-		}
-		rsendsymkeys = append(rsendsymkeys, key)
-	}
-
-	// get incoming symkeys in byte form from both sides and compare
-	var lrecvsymkeys [][]byte
-	for _, id := range lrecvsymkeyids {
-		var key []byte
-		err = clients[0].Call(&key, "pss_getSymmetricKey", id)
-		if err != nil {
-			t.Fatal(err)
-		}
-		match := false
-		for _, otherkey := range rsendsymkeys {
-			if bytes.Equal(otherkey, key) {
-				match = true
-			}
-		}
-		if !match {
-			t.Fatalf("no match right send for left recv key %s", id)
-		}
-		lrecvsymkeys = append(lrecvsymkeys, key)
-	}
-	var rrecvsymkeys [][]byte
-	for _, id := range rrecvsymkeyids {
-		var key []byte
-		err = clients[1].Call(&key, "pss_getSymmetricKey", id)
-		if err != nil {
-			t.Fatal(err)
-		}
-		match := false
-		for _, otherkey := range lsendsymkeys {
-			if bytes.Equal(otherkey, key) {
-				match = true
-			}
-		}
-		if !match {
-			t.Fatalf("no match left send for right recv key %s", id)
-		}
-		rrecvsymkeys = append(rrecvsymkeys, key)
-	}
-
-	// send new handshake request, should send no keys
-	err = clients[0].Call(nil, "pss_handshake", common.ToHex(rpubkey), hextopic, roaddr, false)
-	if err == nil {
-		t.Fatal("expected full symkey buffer error")
-	}
-
-	// expire one key, send new handshake request
-	err = clients[0].Call(nil, "pss_releaseHandshakeKey", common.ToHex(rpubkey), hextopic, lsendsymkeyids[0])
-	if err != nil {
-		t.Fatalf("release left send key %s fail: %v", lsendsymkeyids[0], err)
-	}
-
-	var newlhsendkeyids []string
-
-	// send new handshake request, should now receive one key
-	// check that it is not in previous right recv key array
-	err = clients[0].Call(&newlhsendkeyids, "pss_handshake", common.ToHex(rpubkey), hextopic, roaddr, true)
-	if err != nil {
-		t.Fatal("handshake send fail: %v", err)
-	} else if len(newlhsendkeyids) > 1 {
-		t.Fatalf("wrong receive count, expected 1, got %d", len(newlhsendkeyids))
-	}
-
-	var newlrecvsymkey []byte
-	err = clients[0].Call(&newlrecvsymkey, "pss_getSymmetricKey", newlhsendkeyids[0])
-	if err != nil {
-		t.Fatal(err)
-	}
-	var rmatchsymkeyid *string
-	for i, id := range rrecvsymkeyids {
-		var key []byte
-		err = clients[1].Call(&key, "pss_getSymmetricKey", id)
-		if err != nil {
-			t.Fatal(err)
-		}
-		if bytes.Equal(newlrecvsymkey, key) {
-			rmatchsymkeyid = &rrecvsymkeyids[i]
-		}
-	}
-	if rmatchsymkeyid != nil {
-		t.Fatalf("right sent old key id %s in second handshake", *rmatchsymkeyid)
-	}
-
-}
-
 // send symmetrically encrypted message between two directly connected peers
 func TestSymSend(t *testing.T) {
 	t.Run("32", testSymSend)
@@ -772,117 +544,6 @@ func testAsymSend(t *testing.T) {
 	case cerr := <-lctx.Done():
 		t.Fatalf("test message timed out: %v", cerr)
 	}
-}
-
-// simple ping pong protocol test for the pss devp2p emulation
-func TestProtocol(t *testing.T) {
-	t.Run("32", testProtocol)
-	t.Run("8", testProtocol)
-	t.Run("0", testProtocol)
-}
-
-func testProtocol(t *testing.T) {
-
-	// address hint size
-	var addrsize int64
-	var err error
-	paramstring := strings.Split(t.Name(), "/")
-	addrsize, _ = strconv.ParseInt(paramstring[1], 10, 0)
-	log.Info("protocol test", "addrsize", addrsize)
-
-	hextopic := fmt.Sprintf("%x", PingTopic)
-	clients, err := setupNetwork(2)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	var loaddr []byte
-	err = clients[0].Call(&loaddr, "pss_baseAddr")
-	if err != nil {
-		t.Fatalf("rpc get node 1 baseaddr fail: %v", err)
-	}
-	loaddr = loaddr[:addrsize]
-	var roaddr []byte
-	err = clients[1].Call(&roaddr, "pss_baseAddr")
-	if err != nil {
-		t.Fatalf("rpc get node 2 baseaddr fail: %v", err)
-	}
-	roaddr = roaddr[:addrsize]
-	lnodeinfo := &p2p.NodeInfo{}
-	err = clients[0].Call(&lnodeinfo, "admin_nodeInfo")
-	if err != nil {
-		t.Fatalf("rpc nodeinfo node 11 fail: %v", err)
-	}
-
-	lpubkey := make([]byte, 32)
-	err = clients[0].Call(&lpubkey, "pss_getPublicKey")
-	if err != nil {
-		t.Fatalf("rpc get node 1 pubkey fail: %v", err)
-	}
-	rpubkey := make([]byte, 32)
-	err = clients[1].Call(&rpubkey, "pss_getPublicKey")
-	if err != nil {
-		t.Fatalf("rpc get node 2 pubkey fail: %v", err)
-	}
-
-	time.Sleep(time.Millisecond * 500) // replace with hive healthy code
-
-	lmsgC := make(chan APIMsg)
-	lctx, _ := context.WithTimeout(context.Background(), time.Second*10)
-	lsub, err := clients[0].Subscribe(lctx, "pss", lmsgC, "receive", hextopic)
-	log.Trace("lsub", "id", lsub)
-	defer lsub.Unsubscribe()
-	rmsgC := make(chan APIMsg)
-	rctx, _ := context.WithTimeout(context.Background(), time.Second*10)
-	rsub, err := clients[1].Subscribe(rctx, "pss", rmsgC, "receive", hextopic)
-	log.Trace("rsub", "id", rsub)
-	defer rsub.Unsubscribe()
-
-	// set reciprocal public keys
-	err = clients[0].Call(nil, "pss_setPeerPublicKey", rpubkey, hextopic, roaddr)
-	if err != nil {
-		t.Fatal(err)
-	}
-	err = clients[1].Call(nil, "pss_setPeerPublicKey", lpubkey, hextopic, loaddr)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	// add right peer's public key as protocol peer on left
-	nid, _ := discover.HexID("0x00") // this hack is needed to satisfy the p2p method
-	p := p2p.NewPeer(nid, fmt.Sprintf("%x", loaddr), []p2p.Cap{})
-	pssprotocols[lnodeinfo.ID].protocol.AddPeer(p, pssprotocols[lnodeinfo.ID].run, PingTopic, true, common.ToHex(rpubkey))
-
-	// sends ping asym, checks delivery
-	pssprotocols[lnodeinfo.ID].C <- false
-	select {
-	case <-lmsgC:
-		log.Debug("lnode ok")
-	case cerr := <-lctx.Done():
-		t.Fatalf("test message timed out: %v", cerr)
-	}
-	select {
-	case <-rmsgC:
-		log.Debug("rnode ok")
-	case cerr := <-lctx.Done():
-		t.Fatalf("test message timed out: %v", cerr)
-	}
-
-	// sends ping asym, checks delivery
-	pssprotocols[lnodeinfo.ID].C <- false
-	select {
-	case <-lmsgC:
-		log.Debug("lnode ok")
-	case cerr := <-lctx.Done():
-		t.Fatalf("test message timed out: %v", cerr)
-	}
-	select {
-	case <-rmsgC:
-		log.Debug("rnode ok")
-	case cerr := <-lctx.Done():
-		t.Fatalf("test message timed out: %v", cerr)
-	}
-
 }
 
 // symmetric send performance with varying message sizes
